refactor: move RequestsRoutingPage
