+++ web-client/src/modules/requests/constants.tsx
@@ -3,7 +3,7 @@ import Location from 'react-app-location';
 /*
  * :type is either 'ask' or 'offer'
  */
-export const TabbedRequestsLocation = new Location('/requests/main/:type');
+export const TabbedRequestsLocation = new Location('/requests/list/:type');
 export const AcceptedRequestsLocation = new Location('/requests/accepted');
 export const FindRequestsLocation = new Location('/requests/find');
 export const NewRequestsLocation = new Location('/requests/new');
